+++ packages/icons-flutter/lib/icons.dart
@@ -3,13 +3,13 @@
 /// 2700+ handcrafted, pixel-perfect SVG icons in Outline and Filled weights.
 ///
 /// ```dart
-/// import 'package:vezham_icons_flutter/vezham_icons_flutter.dart';
+/// import 'package:vezham_icons/icons.dart';
 ///
 /// Vezham.outline.home;
 /// Vezham.filled.settings;
 /// vezhamIconSvg(Vezham.outline.star, size: 32, color: '#d97757');
 /// ```
-library vezham_icons_flutter;
+library vezham_icons.icons;
 
 export 'src/icons.dart';
 export 'src/vezham_icon.dart';